motion-recognization/pyproject.toml
[project]
name = "motion-recognization"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mediapipe>=0.10.21",
    "pandas>=2.3.2",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.2",
    "scipy>=1.15.3",
]
